Should be a stable version =) At the beginning of the game you should enter number of mines that will be set on a field. To make a move follow the pattern 'X' 'Y' 'command', where: 'X' - the X-coordinate; 'Y' - the Y-coordinate; 'command' - you can choose between 'free' and 'mine'.
'free' - explores a cell on the field.
'mine' - marks a cell as a mine.

There are 2 ways to win: the first one is to mark all the mines, and the second one is to open all non-mined cells.
All the rules are the same as rules of classic minesweeper game.
Width and height are hardcoded to 9. Maybe it'll be changed sometime ^_^